MSTests/CustomerRepositoryTest.cs
using Microsoft.VisualStudio.TestTools.UnitTesting;
using TravelPort.BL;

namespace MSTests
{
    [TestClass]
    public class CustomerRepositoryTest
    {
        [TestMethod]
        public void RetrieveExisting()
        {
            // Arrange
            var customerRepository = new CustomerRepository();
            var expected = new Customer(1)
            {
                Contact = "test@gmail.com",
                FirstName = "testFirstName",
                LastName = "testLastName"
            };

            // Act
            var actual = customerRepository.Retrieve(1);

            // Assert
            Assert.AreEqual(expected.CustomerId, actual.CustomerId);
            Assert.AreEqual(expected.Contact, actual.Contact);
            Assert.AreEqual(expected.FirstName, actual.FirstName);
            Assert.AreEqual(expected.LastName, actual.LastName);
        }
        
    }
}
